コントローラーを修正

+++ app/controllers/items_controller.rb
@@ -37,8 +37,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.user_id == current_user.id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
